requirements.txt
requests
psycopg2
pandas
spotipy
python-dotenv
sqlalchemy